+++ app/data_models/__init__.py
@@ -5,3 +5,4 @@ from .checklist import Checklist
 from .spot import Spot
 from .plan_spot_map import PlanSpotMap
 from .plan_spot_tag_map import PlanSpotTagMap
+from .administrative_division import AdministrativeDivision
