+++ app/data_models/plan.py
@@ -1,23 +1,25 @@
-import os
-import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-from app.repository.db import Base
 
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime
-from sqlalchemy.orm import relationship
 
-class Plan(Base):
-    __tablename__ = 'plan'
-    
-    plan_id = Column(Integer, primary_key=True, autoincrement=True)
-    plan_name = Column(String(255), nullable=True)
-    start_date = Column(DateTime, nullable=True)
-    end_date = Column(DateTime, nullable=True)
-    main_location = Column(String(50), nullable=True)
-    ages = Column(Integer, nullable=True)
-    companion_count = Column(Integer, nullable=True)
-    plan_concepts = Column(String(255), nullable=True)
-    member_id = Column(Integer, ForeignKey('member.member_id'), nullable=False)
+from datetime import datetime
+from typing import List, Optional
 
-    member = relationship("Member", back_populates="plans")
-    checklist = relationship("Checklist", back_populates="plan")
+from pydantic import Field
+from sqlmodel import Relationship
+
+from travelPlaner_BackEnd.app.data_models.member import Member
+
+
+class Plan(SQLModel, table=True):
+    plan_id: Optional[int] = Field(default=None, primary_key=True)
+    plan_name: Optional[str] = Field(default=None, sa_column_kwargs={"length": 255})
+    start_date: Optional[datetime] = None
+    end_date: Optional[datetime] = None
+    main_location: Optional[str] = Field(default=None, sa_column_kwargs={"length": 50})
+    ages: Optional[int] = None
+    companion_count: Optional[int] = None
+    plan_concepts: Optional[str] = Field(default=None, sa_column_kwargs={"length": 255})
+    member_id: int = Field(foreign_key="member.member_id")
+
+    member: Member = Relationship(back_populates="plans")
+    checklist: Optional["Checklist"] = Relationship(back_populates="plan")
+    plan_spots: List["PlanSpotMap"] = Relationship(back_populates="plan")
