+++ app/data_models/plan_spot_map.py
@@ -1,15 +1,18 @@
-import os
-import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-from app.repository.db import Base
+from typing import Optional
+from pydantic import Field
+from sqlmodel import Relationship
+from datetime import time
+
+from travelPlaner_BackEnd.app.data_models.plan import Plan
+from travelPlaner_BackEnd.app.data_models.spot import Spot
 
 
-from sqlalchemy import Column, Integer, ForeignKey
+class PlanSpotMap(SQLModel, table=True):
+    plan_id: int = Field(foreign_key="plan.plan_id", primary_key=True)
+    spot_id: int = Field(foreign_key="spot.spot_id", primary_key=True)
+    day_x: int = Field(...)
+    order: int = Field(...)
+    time: Optional[time] = Field(default=None)  # 시간 필드 추가
 
-class PlanSpotMap(Base):
-    __tablename__ = 'plan_spot_map'
-    
-    plan_id = Column(Integer, ForeignKey('plan.plan_id'), primary_key=True)
-    spot_id = Column(Integer, ForeignKey('spot.spot_id'), primary_key=True)
-    day_x = Column(Integer, nullable=False)
-    order = Column(Integer, nullable=False)
+    plan: Plan = Relationship(back_populates="plan_spots")
+    spot: Spot = Relationship(back_populates="plan_spots")
