+++ app/data_models/spot.py
@@ -1,28 +1,28 @@
-import os
-import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-from app.repository.db import Base
+from datetime import datetime
+from typing import List, Optional
 
-from sqlalchemy import Column, Integer, String, Boolean, DateTime
-from sqlalchemy.orm import relationship
+from pydantic import Field
+from sqlmodel import Relationship
 
-class Spot(Base):
-    __tablename__ = 'spot'
-    
-    spot_id = Column(Integer, primary_key=True, autoincrement=True)
-    kor_name = Column(String(255), nullable=False)
-    eng_name = Column(String(255), nullable=True)
-    description = Column(String(255), nullable=False)
-    address = Column(String(255), nullable=False)
-    zip = Column(CHAR(10), nullable=False)
-    url = Column(String(2083), nullable=True)
-    image_url = Column(String(2083), nullable=False)
-    map_url = Column(String(2083), nullable=False)
-    likes = Column(Integer, nullable=True)
-    satisfaction = Column(Float, nullable=True)
-    created_at = Column(DateTime, nullable=False)
-    updated_at = Column(DateTime, nullable=False)
-    spot_category = Column(Integer, nullable=False)
-    phone_number = Column(String(300), nullable=True)
-    business_status = Column(Boolean, nullable=True)
-    business_hours = Column(String(255), nullable=True)
+
+class Spot(SQLModel, table=True):
+    spot_id: Optional[int] = Field(default=None, primary_key=True)
+    kor_name: str = Field(sa_column_kwargs={"length": 255})
+    eng_name: Optional[str] = Field(default=None, sa_column_kwargs={"length": 255})
+    description: str = Field(sa_column_kwargs={"length": 255})
+    address: str = Field(sa_column_kwargs={"length": 255})
+    zip: str = Field(sa_column_kwargs={"length": 10})
+    url: Optional[str] = Field(default=None, sa_column_kwargs={"length": 2083})
+    image_url: str = Field(sa_column_kwargs={"length": 2083})
+    map_url: str = Field(sa_column_kwargs={"length": 2083})
+    likes: Optional[int] = None
+    satisfaction: Optional[float] = None
+    created_at: datetime
+    updated_at: datetime
+    spot_category: int
+    phone_number: Optional[str] = Field(default=None, sa_column_kwargs={"length": 300})
+    business_status: Optional[bool] = None
+    business_hours: Optional[str] = Field(default=None, sa_column_kwargs={"length": 255})
+
+    plan_spots: List["PlanSpotMap"] = Relationship(back_populates="spot")
+    spot_tags: List["PlanSpotTagMap"] = Relationship(back_populates="spot")
