--- app/schema/__init__.py
@@ -1,17 +1,0 @@
-from .member import Member
-from .plan import Plan
-from .spot_tag import SpotTag
-from .checklist import Checklist
-from .spot import Spot
-from .plan_spot_map import PlanSpotMap
-from .plan_spot_tag_map import PlanSpotTagMap
-
-__all__ = [
-    "Member",
-    "Plan",
-    "SpotTag",
-    "Checklist",
-    "Spot",
-    "PlanSpotMap",
-    "PlanSpotTagMap",
-]
